Change margin for banner and transfer code
Changed the #nav-menu position to relative, so it would not cover the main content. Changed the margin for the banner and the search bar in the index page to keep the distance between each other. Transferred the code from the index to the script.js, as it will be used for all pages.