Add code to change navbar according to viewport
Added a code that changes the navbar according to the position of the viewport. If the viewport pass the bottom of the header, the navbar changes and the one with the menu is "hidden".The menu button still works, and when the header becomes visible again, the navabar changes back.
Also, when the page is loaded, it checks the width to determine whether the mobile navbar should be hidden or not.