Add links between page and fixing problems
Added "a" tags to connect the HTML pages through the correct elements; changed the comments that distinguish the main content of the pages catalogo.html and produto.html, since they wrong. Also added a timeout for the focusout event on the navitems, because whenever a suboption was clicked, they would disappear before going to the correct page.